requirements.txt
asgiref==3.5.2
Babel==2.10.3
boto==2.49.0
boto3==1.24.57
botocore==1.27.57
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-autoslug==1.9.8
django-cities-light==3.9.1
django-ckeditor==6.5.0
django-countries==7.3.2
django-country==0.1.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-money==3.0.0
django-phonenumber-field==6.3.0
django-smart-selects==1.6.0
django-storages==1.13.1
django-taggit==3.0.0
idna==3.3
jmespath==1.0.1
lxml==4.9.1
oauthlib==3.2.0
phonenumbers==8.12.52
Pillow==9.2.0
progressbar2==4.0.0
psycopg2==2.9.3
py-moneyed==2.0
pycairo==1.21.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-utils==3.3.3
python3-openid==3.2.0
pytz==2022.1
rave-python==1.2.16
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
typing_extensions==4.3.0
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.10
webencodings==0.5.1
whitenoise==6.2.0
